requirements.txt
Django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
# mysql-connector-python==8.0.31
# mysqlclient==2.1.1
redis==3.5.3
django-redis==5.3.0
django-q==1.3.9
docker==5.0.3
urllib3==1.26.4